Move GasPrice and GasUnit to separate submodules
